requirements.txt
pyspark
kafka-python
faker
pydantic
python-dotenv
scikit-learn
xgboost
lightgbm
networkx
pandas
pyarrow
fastapi
uvicorn
streamlit
